+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inbox/CTIconMessageViewHolder.java
@@ -105,7 +105,7 @@ class CTIconMessageViewHolder extends CTInboxBaseMessageViewHolder {
                         hideTwoButtons(this.cta1, this.cta2, this.cta3);
                         if (parentWeak != null) {
                             this.cta1.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta1.getText().toString(), cta1Object, parentWeak));
+                                    this.cta1.getText().toString(), cta1Object, parentWeak,false));
                         }
                         break;
                     case 2:
@@ -122,9 +122,9 @@ class CTIconMessageViewHolder extends CTInboxBaseMessageViewHolder {
                         hideOneButton(this.cta1, this.cta2, this.cta3);
                         if (parentWeak != null) {
                             this.cta1.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta1.getText().toString(), cta1Object, parentWeak));
+                                    this.cta1.getText().toString(), cta1Object, parentWeak,false));
                             this.cta2.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta2.getText().toString(), cta2Object, parentWeak));
+                                    this.cta2.getText().toString(), cta2Object, parentWeak,false));
                         }
                         break;
                     case 3:
@@ -145,11 +145,11 @@ class CTIconMessageViewHolder extends CTInboxBaseMessageViewHolder {
                         this.cta3.setBackgroundColor(Color.parseColor(content.getLinkBGColor(cta3Object)));
                         if (parentWeak != null) {
                             this.cta1.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta1.getText().toString(), cta1Object, parentWeak));
+                                    this.cta1.getText().toString(), cta1Object, parentWeak,false));
                             this.cta2.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta2.getText().toString(), cta2Object, parentWeak));
+                                    this.cta2.getText().toString(), cta2Object, parentWeak,false));
                             this.cta3.setOnClickListener(new CTInboxButtonClickListener(position, inboxMessage,
-                                    this.cta3.getText().toString(), cta3Object, parentWeak));
+                                    this.cta3.getText().toString(), cta3Object, parentWeak,false));
                         }
                         break;
                 }
@@ -387,7 +387,6 @@ class CTIconMessageViewHolder extends CTInboxBaseMessageViewHolder {
                             if (readDot.getVisibility() == View.VISIBLE) {
                                 parent.didShow(null, position);
                             }
-                            readDot.setVisibility(View.GONE);
                         }
                     });
                 }
@@ -422,7 +421,7 @@ class CTIconMessageViewHolder extends CTInboxBaseMessageViewHolder {
 
         if (parentWeak != null) {
             clickLayout.setOnClickListener(
-                    new CTInboxButtonClickListener(position, inboxMessage, null, null, parentWeak));
+                    new CTInboxButtonClickListener(position, inboxMessage, null, null, parentWeak,true));
         }
     }
 }