+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inbox/CTInboxListViewFragment.java
@@ -42,7 +42,7 @@ public class CTInboxListViewFragment extends Fragment {
     interface InboxListener {
 
         void messageDidClick(Context baseContext, CTInboxMessage inboxMessage, Bundle data,
-                HashMap<String, String> keyValue);
+                HashMap<String, String> keyValue,boolean isBodyClick);
 
         void messageDidShow(Context baseContext, CTInboxMessage inboxMessage, Bundle data);
     }
@@ -58,6 +58,8 @@ public class CTInboxListViewFragment extends Fragment {
     MediaPlayerRecyclerView mediaRecyclerView;
 
     RecyclerView recyclerView;
+    private  CTInboxMessageAdapter inboxMessageAdapter;
+
 
     CTInboxStyleConfig styleConfig;
 
@@ -75,16 +77,28 @@ public class CTInboxListViewFragment extends Fragment {
             config = bundle.getParcelable("config");
             styleConfig = bundle.getParcelable("styleConfig");
             tabPosition = bundle.getInt("position", -1);
-            final String filter = bundle.getString("filter", null);
+            updateInboxMessages();
             if (context instanceof CTInboxActivity) {
                 setListener((CTInboxListViewFragment.InboxListener) getActivity());
             }
-            CleverTapAPI cleverTapAPI = CleverTapAPI.instanceWithConfig(getActivity(), config);
-            if (cleverTapAPI != null) {
-                ArrayList<CTInboxMessage> allMessages = cleverTapAPI.getAllInboxMessages();
-                inboxMessages = filter != null ? filterMessages(allMessages, filter) : allMessages;
-            }
         }
+    }
+    private void updateInboxMessages(){
+        Bundle bundle = getArguments();
+        if(bundle==null) return;
+        final String filter = bundle.getString("filter", null);
+        CleverTapAPI cleverTapAPI = CleverTapAPI.instanceWithConfig(getActivity(), config);
+        if (cleverTapAPI != null) {
+            Logger.v( "CTInboxListViewFragment:onAttach() called with: tabPosition = [" + tabPosition + "], filter = [" + filter + "]");
+            ArrayList<CTInboxMessage> allMessages = cleverTapAPI.getAllInboxMessages();
+            inboxMessages = filter != null ? filterMessages(allMessages, filter) : allMessages;
+        }
+    }
+
+    void updateAdapterContent(){
+        updateInboxMessages();
+        if(inboxMessageAdapter==null || inboxMessages==null || config ==null ) return;
+        inboxMessageAdapter.updateInboxMessages(inboxMessages);
     }
 
     @Nullable
@@ -106,7 +120,7 @@ public class CTInboxListViewFragment extends Fragment {
         noMessageView.setVisibility(View.GONE);
 
         final LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
-        final CTInboxMessageAdapter inboxMessageAdapter = new CTInboxMessageAdapter(inboxMessages, this);
+        inboxMessageAdapter= new CTInboxMessageAdapter(inboxMessages, this);
 
         if (haveVideoPlayerSupport) {
             mediaRecyclerView = new MediaPlayerRecyclerView(getActivity());
@@ -202,12 +216,11 @@ public class CTInboxListViewFragment extends Fragment {
         }
     }
 
-    void didClick(Bundle data, int position, HashMap<String, String> keyValuePayload) {
+    void didClick(Bundle data, int position, HashMap<String, String> keyValuePayload, boolean isInboxMessageBodyClick) {
         CTInboxListViewFragment.InboxListener listener = getListener();
         if (listener != null) {
             //noinspection ConstantConditions
-            listener.messageDidClick(getActivity().getBaseContext(), inboxMessages.get(position), data,
-                    keyValuePayload);
+            listener.messageDidClick(getActivity().getBaseContext(), inboxMessages.get(position), data, keyValuePayload, isInboxMessageBodyClick);
         }
     }
 
@@ -215,6 +228,7 @@ public class CTInboxListViewFragment extends Fragment {
     void didShow(Bundle data, int position) {
         CTInboxListViewFragment.InboxListener listener = getListener();
         if (listener != null) {
+            Logger.v("CTInboxListViewFragment:didShow() called with: data = [" + data + "], position = [" + position + "]");
             //noinspection ConstantConditions
             listener.messageDidShow(getActivity().getBaseContext(), inboxMessages.get(position), data);
         }
@@ -257,8 +271,7 @@ public class CTInboxListViewFragment extends Fragment {
         this.mediaRecyclerView = mediaRecyclerView;
     }
 
-    void handleClick(int position, String buttonText, JSONObject jsonObject,
-            HashMap<String, String> keyValuePayload) {
+    void handleClick(int position, String buttonText, JSONObject jsonObject, HashMap<String, String> keyValuePayload, boolean isInboxMessageBodyClick) {
         try {
             Bundle data = new Bundle();
             JSONObject wzrkParams = inboxMessages.get(position).getWzrkParams();
@@ -273,7 +286,7 @@ public class CTInboxListViewFragment extends Fragment {
             if (buttonText != null && !buttonText.isEmpty()) {
                 data.putString("wzrk_c2a", buttonText);
             }
-            didClick(data, position, keyValuePayload);
+            didClick(data, position, keyValuePayload,isInboxMessageBodyClick);
             boolean isKVButton = keyValuePayload != null && !keyValuePayload.isEmpty();
             if (jsonObject != null) {
                 if (isKVButton || inboxMessages.get(position).getInboxMessageContents().get(0).getLinktype(jsonObject)
@@ -298,7 +311,7 @@ public class CTInboxListViewFragment extends Fragment {
         }
     }
 
-    void handleViewPagerClick(int position, int viewPagerPosition) {
+    void handleViewPagerClick(int position, int viewPagerPosition,boolean isInboxMessageBodyClick) {
         try {
             Bundle data = new Bundle();
             JSONObject wzrkParams = inboxMessages.get(position).getWzrkParams();
@@ -309,7 +322,7 @@ public class CTInboxListViewFragment extends Fragment {
                     data.putString(keyName, wzrkParams.getString(keyName));
                 }
             }
-            didClick(data, position, null);
+            didClick(data, position, null,isInboxMessageBodyClick);
             String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(viewPagerPosition)
                     .getActionUrl();
             fireUrlThroughIntent(actionUrl);