+++ clevertap-core/src/main/java/com/clevertap/android/sdk/pushnotification/PushConstants.java
@@ -3,6 +3,7 @@ package com.clevertap.android.sdk.pushnotification;
 import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
 import androidx.annotation.StringDef;
+import com.clevertap.android.sdk.Logger;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
@@ -47,11 +48,11 @@ public interface PushConstants {
     }
 
     enum PushType {
-        FCM(FCM_DELIVERY_TYPE, FCM_PROPERTY_REG_ID, CT_FIREBASE_PROVIDER_CLASS, FIREBASE_SDK_CLASS, ALL_DEVICES),
-        XPS(XIAOMI_DELIVERY_TYPE, XPS_PROPERTY_REG_ID, CT_XIAOMI_PROVIDER_CLASS, XIAOMI_SDK_CLASS, ALL_DEVICES),
-        HPS(HMS_DELIVERY_TYPE, HPS_PROPERTY_REG_ID, CT_HUAWEI_PROVIDER_CLASS, HUAWEI_SDK_CLASS, ALL_DEVICES),
-        BPS(BAIDU_DELIVERY_TYPE, BPS_PROPERTY_REG_ID, CT_BAIDU_PROVIDER_CLASS, BAIDU_SDK_CLASS, ALL_DEVICES),
-        ADM(ADM_DELIVERY_TYPE, ADM_PROPERTY_REG_ID, CT_ADM_PROVIDER_CLASS, ADM_SDK_CLASS, ALL_DEVICES);
+        FCM(FCM_DELIVERY_TYPE, FCM_PROPERTY_REG_ID, CT_FIREBASE_PROVIDER_CLASS, FIREBASE_SDK_CLASS, ALL_DEVICES, DEFAULT_PUSH_TYPE_REGION),
+        XPS(XIAOMI_DELIVERY_TYPE, XPS_PROPERTY_REG_ID, CT_XIAOMI_PROVIDER_CLASS, XIAOMI_SDK_CLASS, ALL_DEVICES, DEFAULT_PUSH_TYPE_REGION),
+        HPS(HMS_DELIVERY_TYPE, HPS_PROPERTY_REG_ID, CT_HUAWEI_PROVIDER_CLASS, HUAWEI_SDK_CLASS, ALL_DEVICES, DEFAULT_PUSH_TYPE_REGION),
+        BPS(BAIDU_DELIVERY_TYPE, BPS_PROPERTY_REG_ID, CT_BAIDU_PROVIDER_CLASS, BAIDU_SDK_CLASS, ALL_DEVICES, DEFAULT_PUSH_TYPE_REGION),
+        ADM(ADM_DELIVERY_TYPE, ADM_PROPERTY_REG_ID, CT_ADM_PROVIDER_CLASS, ADM_SDK_CLASS, ALL_DEVICES, DEFAULT_PUSH_TYPE_REGION);
 
         private final String ctProviderClassName;
 
@@ -61,16 +62,19 @@ public interface PushConstants {
 
         private final String type;
 
+        private String serverRegion ;
+
         private @XiaomiPush
         int runningDevices;
 
         PushType(@DeliveryType String type, @RegKeyType String prefKey, @CTPushProviderClass String className,
-                @PushMessagingClass String messagingSDKClassName, @XiaomiPush int runningDevices) {
+                @PushMessagingClass String messagingSDKClassName, @XiaomiPush int runningDevices, String region) {
             this.type = type;
             this.tokenPrefKey = prefKey;
             this.ctProviderClassName = className;
             this.messagingSDKClassName = messagingSDKClassName;
             this.runningDevices = runningDevices;
+            this.serverRegion = region;
         }
 
         public String getCtProviderClassName() {
@@ -96,6 +100,26 @@ public interface PushConstants {
 
         public @XiaomiPush int getRunningDevices() {
             return runningDevices;
+        }
+
+
+        /**
+         * Sets the Server Room Region for various push providers. This info will be passed onto
+         * CleverTap Servers which will then access the 3rd party push providers from appropriate
+         * region to push notifications.
+         * @param region   The Server Room region
+         */
+        public void  setServerRegion(@NonNull String region){
+            Logger.v("PushConstants: setServerRegion called with region:"+region);
+
+            this.serverRegion = region;
+        }
+        /**
+         * Get Server Room Region for various push providers
+         */
+        public String getServerRegion(){
+            Logger.v("PushConstants: getServerRegion called, returning region:"+serverRegion);
+            return  this.serverRegion;
         }
 
         @NonNull
@@ -139,6 +163,7 @@ public interface PushConstants {
     String BPS_PROPERTY_REG_ID = "bps_token";
     String HPS_PROPERTY_REG_ID = "hps_token";
     String ADM_PROPERTY_REG_ID = "adm_token";
+    String DEFAULT_PUSH_TYPE_REGION = "";
     /**
      * Android platform type. Only GCM transport will be allowed.
      */