+++ sample/src/main/java/com/clevertap/demo/MyFcmMessageListenerService.kt
@@ -1,5 +1,7 @@
 package com.clevertap.demo
 
+import android.os.Looper
+import com.clevertap.android.sdk.CleverTapAPI
 import com.clevertap.android.sdk.pushnotification.fcm.CTFcmMessageHandler
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
@@ -8,10 +10,14 @@ class MyFcmMessageListenerService : FirebaseMessagingService() {
 
     override fun onMessageReceived(message: RemoteMessage) {
         super.onMessageReceived(message)
+        message.data["nh_source"] = "MyFcmMessageListenerService"
         var pushType = "fcm"
         if (pushType.equals("fcm")) {
-            CTFcmMessageHandler()
-                .createNotification(applicationContext, message)
+            android.os.Handler(Looper.getMainLooper()).post {
+                println("MyFcmMessageListenerService onMessageReceived createNotification on ${Thread.currentThread()}")
+                CTFcmMessageHandler()
+                    .createNotification(applicationContext, message)
+            }
             //CTFcmMessageHandler().processPushAmp(applicationContext, message)
         } else if (pushType.equals("hps")) {
             //CTHmsMessageHandler().createNotification(applicationContext,message)
@@ -19,6 +25,13 @@ class MyFcmMessageListenerService : FirebaseMessagingService() {
         } else if (pushType.equals("xps")) {
             //CTXiaomiMessageHandler().createNotification(applicationContext,message)
             //CTXiaomiMessageHandler().processPushAmp(applicationContext,message)
+        }
+    }
+
+    override fun onNewToken(token: String) {
+        super.onNewToken(token)
+        CleverTapAPI.getDefaultInstance(this)?.apply {
+            pushFcmRegistrationId(token, true)
         }
     }
 }